Fix refresh token fail detection
