+++ app/src/main/java/com/yara/raco/workers/RefreshTokenWorker.kt
@@ -12,6 +12,17 @@ class RefreshTokenWorker(context: Context, workerParams: WorkerParameters) :
     CoroutineWorker(context, workerParams) {
     override suspend fun doWork(): Result {
         return withContext(Dispatchers.IO) {
+            if (getLastExecutionResult(applicationContext) == ResultCode.INVALID_TOKEN) {
+                return@withContext Result.failure(
+                    workDataOf(
+                        Pair(
+                            "RESULT_CODE",
+                            ResultCode.INVALID_TOKEN
+                        )
+                    )
+                )
+            }
+
             val userController = UserController.getInstance(applicationContext)
 
             if (!userController.isLoggedIn) {
@@ -20,7 +31,14 @@ class RefreshTokenWorker(context: Context, workerParams: WorkerParameters) :
 
             when (userController.refreshToken()) {
                 ResultCode.SUCCESS -> return@withContext Result.success()
-                ResultCode.INVALID_TOKEN -> return@withContext Result.failure()
+                ResultCode.INVALID_TOKEN -> return@withContext Result.failure(
+                    workDataOf(
+                        Pair(
+                            "RESULT_CODE",
+                            ResultCode.INVALID_TOKEN
+                        )
+                    )
+                )
                 else -> return@withContext Result.retry()
             }
         }
@@ -50,6 +68,26 @@ class RefreshTokenWorker(context: Context, workerParams: WorkerParameters) :
                 ExistingPeriodicWorkPolicy.KEEP,
                 getOnWorkRequest()
             )
+        }
+
+        fun getLastExecutionResult(context: Context): Int {
+            val workInfos = WorkManager.getInstance(context)
+                .getWorkInfosByTag("com.yara.raco.REFRESH_TOKEN_WORK").get()
+            val lastWork = workInfos.lastOrNull {
+                it.state != WorkInfo.State.RUNNING && it.state != WorkInfo.State.ENQUEUED
+            } ?: return ResultCode.SUCCESS
+
+            return if (lastWork.state == WorkInfo.State.SUCCEEDED) {
+                ResultCode.SUCCESS
+            } else {
+                if (lastWork.outputData.getInt(
+                        "RESULT_CODE",
+                        ResultCode.UNKNOWN
+                    ) == ResultCode.INVALID_TOKEN
+                ) {
+                    ResultCode.INVALID_TOKEN
+                } else ResultCode.UNKNOWN
+            }
         }
     }
 }
